Changed analyze and bvolume IO routines to assume that the TR on disk is in seconds and so convert the TR to msec as it is read in or back to sec as it is read out. This is needed as part of a change to DICOMRead.c in which the TR for mosaics is computed in msec to be consistent with non-mosaics.
Changed NIFTI1 and NII routines to save TR in sec as a default. In
theory, it should not matter whether it is saved in sec or msec as
long as the xyzt units parameter is set properly. However, I suspect
that there will be many applications out there that won't read the
units parameter and just assume that the TR is in sec.